Добавил возможность оставить комментарий к видео + лайк/убрать лайк

+++ app/src/main/java/com/example/vcontachim/models/VideoComment.kt
@@ -23,7 +23,7 @@ data class ItemVideoComment(
 
 data class LikesVideoComment(
     @SerializedName("user_likes")
-    val userLikes: Long
+    val userLikes: Int
 )
 
 data class ProfileVideoComment(
